# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "hw7-swagger",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This API provides CRUD operations for managing contacts and supports user registration and login. \nUsers can create, view, update, and delete their contacts after authentication. The API uses JWT for authentication.\n"
  },
  "servers": [
    {
      "url": "http://localhost:300",
      "description": "Local development server"
    },
    {
      "url": "https://nodejs-hw-mongodb-4t6o.onrender.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/contacts/{contactId}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "operationId": "getContactById",
        "description": "Returns a contact by its unique ID.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "description": "Unique ID of the contact",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "message": { "type": "string", "example": "Successfully found contact with id 68725a5e1785c8943512c751!" },
                    "data": { "$ref": "#/components/schemas/Contact" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "message": { "type": "string", "example": "Contact not found" }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Update a contact by ID",
        "operationId": "updateContactById",
        "description": "Updates the contact by its unique ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "description": "Unique ID of the contact",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ContactInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "message": { "type": "string", "example": "Contact with id 68725a5e1785c8943512c751 updated successfully!" },
                    "data": { "$ref": "#/components/schemas/Contact" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "message": { "type": "string", "example": "No data provided for update" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "message": { "type": "string", "example": "Contact not found" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete a contact by ID",
        "operationId": "deleteContactById",
        "description": "Deletes the contact by its unique ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "description": "Unique ID of the contact",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted successfully"
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "error" },
                    "message": { "type": "string", "example": "Contact not found" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "64cde2b1a3bbd14d9b2c5ad2" },
          "name": { "type": "string", "example": "Jane Doe" },
          "email": { "type": "string", "example": "jane@mail.com" },
          "phone": { "type": "string", "example": "+380991234567" }
        },
        "required": ["_id", "name", "email", "phone"]
      },
      "ContactInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Jane Doe" },
          "email": { "type": "string", "example": "jane@mail.com" },
          "phone": { "type": "string", "example": "+380991234567" }
        },
        "required": ["name", "email", "phone"]
      },
...
    }
  }
}